feat: add button Download
