+++ src/components/flyout/flyout.module.scss
@@ -1,0 +1,14 @@
+.flyout {
+    position: fixed;
+    padding: 15px;
+    border-radius: 5px;
+    bottom: 10px;
+    left: 10px;
+    z-index: 10;
+    background-color: #cac4c4;
+}
+
+.flyout__button {
+    display: flex;
+    gap: 10px;
+}
