+++ src/page/homePage/homePage.module.scss
@@ -48,3 +48,8 @@
     z-index: 10;
     background-color: #cac4c4;
 }
+
+.flyout__button {
+    display: flex;
+    gap: 10px;
+}
