+++ vitest.config.ts
@@ -10,6 +10,11 @@ export default defineConfig({
     globals: true,
     environment: 'jsdom',
     setupFiles: './vitest-setup.ts',
-    include: ['**/*.test.?(c|m)[jt]s?(x)'],
+    coverage: {
+      provider: 'v8',
+      enabled: true,
+      all: true,
+      reporter: ['text'],
+    },
   },
 });
